+++ src/Command/Plugin/DownloadCommand.php
@@ -3,7 +3,6 @@
 namespace OmekaCli\Command\Plugin;
 
 use GetOptionKit\OptionCollection;
-use OmekaCli\Console\Prompt;
 use OmekaCli\Plugin\Repository\OmekaDotOrgRepository;
 use OmekaCli\Plugin\Repository\GithubRepository;
 
@@ -17,20 +16,20 @@ class DownloadCommand extends AbstractPluginCommand
     public function getUsage()
     {
         return "Usage:\n"
-             . "\tplugin-download [<options>] PLUGIN_NAME\n"
-             . "\tpldl [<options>] PLUGIN_NAME\n"
+             . "\tplugin-download [<options>] <plugin-id>\n"
+             . "\tpldl [<options>] <plugin-id>\n"
+             . "\n"
+             . "<plugin-id> is the identifier of the plugin, as returned by plugin-search\n"
              . "\n"
              . "Options:\n"
              . "\t-f, --force    Force download, even if Omeka minimum version\n"
-             . "\t               requirement is not met\n"
-             . "\t-G, --exclude-github    Do not download plugins from Github\n";
+             . "\t               requirement is not met\n";
     }
 
     public function getOptionsSpec()
     {
         $optionsSpec = new OptionCollection();
         $optionsSpec->add('f|force', 'Force download');
-        $optionsSpec->add('G|exclude-github', 'Do not download plugins from Github');
 
         return $optionsSpec;
     }
@@ -38,45 +37,40 @@ class DownloadCommand extends AbstractPluginCommand
     public function run($options, $args)
     {
         if (count($args) != 1) {
-            $this->logger->error($this->getUsage());
+            $this->logger->error('Bad number of arguments');
+            fwrite(STDERR, $this->getUsage());
 
             return 1;
         }
 
-        $repositories = array();
-        $repositories[] = new OmekaDotOrgRepository();
-        if (!isset($options['exclude-github']) || !$options['exclude-github']) {
-            $repositories[] = new GithubRepository();
-        }
-
-        $pluginName = reset($args);
+        $repositories = array(
+            new OmekaDotOrgRepository(),
+            new GithubRepository(),
+        );
 
         foreach ($repositories as $repository) {
-            $pluginsInfo = $repository->find($pluginName);
-            foreach ($pluginsInfo as $pluginInfo) {
-                $plugins[] = array(
+            $repository->setLogger($this->logger);
+        }
+
+        $id = reset($args);
+
+        $plugin = null;
+        foreach ($repositories as $repository) {
+            $pluginInfo = $repository->find($id);
+            if ($pluginInfo) {
+                $plugin = array(
                     'info' => $pluginInfo,
                     'repository' => $repository,
                 );
+
+                break;
             }
         }
 
-        if (empty($plugins)) {
-            $this->logger->error('plugin not found');
+        if (!isset($plugin)) {
+            $this->logger->error('Plugin not found');
 
             return 1;
-        }
-
-        if (1 === count($plugins)) {
-            $plugin = reset($plugins);
-        } else {
-            $plugin = $this->pluginPrompt($plugins);
-        }
-
-        if (!$plugin) {
-            $this->logger->info('Nothing downloaded');
-
-            return 0;
         }
 
         $destDir = '.';
@@ -97,43 +91,46 @@ class DownloadCommand extends AbstractPluginCommand
         $this->logger->info('Downloading plugin');
         try {
             $repository = $plugin['repository'];
-            $dest = $repository->download($plugin['info'], $destDir);
+            $tempdir = $repository->download($id);
+
+            foreach (scandir($tempdir) as $file) {
+                $matches = array();
+                if (preg_match('/(.*)Plugin.php$/', $file, $matches)) {
+                    $pluginName = $matches[1];
+                    break;
+                }
+            }
+
+            if (!isset($pluginName)) {
+                // Try to guess plugin name from id
+                $parts = preg_split('/[\/-]/', $id);
+                $pluginName = end($parts);
+            }
+
+            $dest = $destDir . '/' . $pluginName;
+            if (file_exists($dest)) {
+                $this->logger->error('Destination already exists : {dest}', array('dest' => $dest));
+
+                return 1;
+            }
+
+            if (false == rename($tempdir, $dest)) {
+                $this->logger->error('Cannot move {tempdir} to {dest}', array(
+                    'tempdir' => $tempdir,
+                    'dest' => $dest,
+                ));
+
+                return 1;
+            }
+
+            chmod($dest, 0755);
             $this->logger->notice('Plugin downloaded into {path}', array('path' => $dest));
         } catch (\Exception $e) {
             $this->logger->error('Download failed: {message}', array('message' => $e->getMessage()));
+
+            return 1;
         }
 
         return 0;
-    }
-
-    protected function pluginPrompt($plugins)
-    {
-        if (empty($plugins)) {
-            return null;
-        }
-
-        $this->logger->info('{count} plugin(s) found', array('count' => count($plugins)));
-
-        foreach ($plugins as $plugin) {
-            $toMenu[] = sprintf('%s (%s) - %s',
-                $plugin['info']['displayName'],
-                $plugin['info']['version'],
-                (array_key_exists('owner', $plugin['info']))
-                    ? $plugin['repository']->getDisplayName()
-                        . '/' . $plugin['info']['owner']
-                    : $plugin['repository']->getDisplayName()
-            );
-        }
-
-        if (isset($toMenu)) {
-            $prompt = new Prompt();
-            $chosenIdx = $prompt->menu('Choose one', $toMenu);
-        }
-
-        if ($chosenIdx >= 0) {
-            $chosenPlugin = $plugins[$chosenIdx];
-        }
-
-        return (isset($chosenPlugin)) ? $chosenPlugin : null;
     }
 }