+++ src/Plugin/Repository/AbstractRepository.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace OmekaCli\Plugin\Repository;
+
+use Psr\Log\LoggerAwareTrait;
+use Psr\Log\NullLogger;
+
+abstract class AbstractRepository implements RepositoryInterface
+{
+    use LoggerAwareTrait;
+
+    public function __construct()
+    {
+        $this->setLogger(new NullLogger());
+    }
+}