+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- New command plugin-search
+
+## Changed
+- Command plugin-download requires an exact name for the plugin
+
+## Removed
+- OmekaCli\Console\Prompt
+
 ## Fixed
 - Phar compiler adds shebang
 - Prevent warnings if command class does not exist